feat: Refactor Buttons
