+++ src/components/index.tsx
@@ -1,8 +1,5 @@
 export * from "./Badge";
 export * from "./Button";
-export * from "./Button/BackButton";
-export * from "./Button/CancelButton";
-export * from "./Button/JoinButton";
 export * from "./Calendar";
 export * from "./Card";
 export * from "./Confirm";
